verify user's id to project's userId for [id] routes

+++ app/api/users/[username]/projects/[id]/route.js
@@ -5,27 +5,25 @@ import {
 } from "@/controllers/projects";
 import { connectDb } from "@/lib/connectDb";
 import { jsonWithCors, optionsWithCors, textWithCors } from "@/lib/cors";
-import { getErrorResponse, denyAccess } from "@/lib/errorHandler";
+import { getErrorResponse, denyAccess, notFound } from "@/lib/errorHandler";
 import { requireAuth } from "@/lib/auth";
+import { getUserByUsername } from "@/controllers/users";
 
 //* /api/projects/[id]
 
-//* GET
+//* GET - public
 export async function GET(req, { params }) {
   try {
     await connectDb();
 
     const { username, id } = await params;
 
-    // require authorization
-    const authUser = requireAuth(req);
+    // vertify ownership of project to current user
     const user = await getUserByUsername(username);
-
-    if (authUser.id !== user.id) {
+    const project = await getProjectById(id);
+    if (project.userId !== user.id) {
       throw denyAccess("No authorization to access this project");
     }
-
-    const project = await getProjectById(id);
 
     return jsonWithCors(project, req);
   } catch (err) {
@@ -34,18 +32,24 @@ export async function GET(req, { params }) {
   }
 }
 
-//* PUT
+//* PUT - private
 export async function PUT(req, { params }) {
   try {
     await connectDb();
 
     const { username, id } = await params;
 
+    // verify username to current user
     const authUser = requireAuth(req);
     const user = await getUserByUsername(username);
-
     if (authUser.id !== user.id) {
       throw denyAccess("No authorization to access this project");
+    }
+
+    // vertify ownership of project to current user
+    const existingProject = await getProjectById(id);
+    if (existingProject.userId !== user.id) {
+      throw notFound("Project not found");
     }
 
     const data = await req.json();
@@ -58,18 +62,24 @@ export async function PUT(req, { params }) {
   }
 }
 
-//* DELETE
+//* DELETE - private
 export async function DELETE(req, { params }) {
   try {
     await connectDb();
 
     const { username, id } = await params;
 
+    // verify username to current user
     const authUser = requireAuth(req);
     const user = await getUserByUsername(username);
-
     if (authUser.id !== user.id) {
       throw denyAccess("No authorization to access this project");
+    }
+
+    // vertify ownership of project to current user
+    const existingProject = await getProjectById(id);
+    if (existingProject.userId !== user.id) {
+      throw notFound("Project not found");
     }
 
     const project = await deleteProject(id);
